Simplify page centering logic and minimap managing
Now perfect-margin--minimap-window-p is able to detect both the
minimap temporary and final window (it does it through the name
of the minimap buffer, as minimap-get-window does too). Before,
only the final window was detected, and not the _temp one.

The ubique minimap detection allows the removal of the function
perfect-margin--minimap-left-adjacent-covered-p.  No longer the
location of the minimap should be a concern to page centering.

The generic perfect-margin--init-window-margins allows to apply
margins to any window which requires it to center its page. The
presence of more windows (minimap included) is handled rightly.

The function perfect-margin-minimap-margin-window is now merged
with the main function perfect-margin-margin-windows. A minimap
temporary or final window is detected without requiring to know
its location, and the use of indirect buffers (the minimap will
have upon its creation the same name of the target window) does
not iterfere with page centering.  Also, only live windows will
be processed.

The unused perfect-margin--width-with-margins function is still
in the code, as an utility function.